refactor: moved fs-router to standalone package
