+++ packages/fs-router-vite-plugin/src/compiler.ts
