+++ packages/fs-router-vite-plugin/src/constants.ts
