+++ packages/fs-router-vite-plugin/src/eliminateUnreferencedIdentifiers.ts
