+++ packages/fs-router-vite-plugin/src/routes-generator/generator.ts
