+++ packages/fs-router-vite-plugin/src/routes-generator/index.ts
