+++ packages/fs-router-vite-plugin/src/routes-generator/types.ts
