+++ packages/fs-router-vite-plugin/src/routes-generator/utils.ts
